Restore test code shared with SparseSeries tests
