+++ pandas/tests/series/test_operators.py
@@ -28,38 +28,6 @@ class TestSeriesOperators(TestData, tm.TestCase):
 
     _multiprocess_can_split_ = True
 
-    def test_binop_maybe_preserve_name(self):
-        # names match, preserve
-        result = self.ts * self.ts
-        self.assertEqual(result.name, self.ts.name)
-        result = self.ts.mul(self.ts)
-        self.assertEqual(result.name, self.ts.name)
-
-        result = self.ts * self.ts[:-2]
-        self.assertEqual(result.name, self.ts.name)
-
-        # names don't match, don't preserve
-        cp = self.ts.copy()
-        cp.name = 'something else'
-        result = self.ts + cp
-        self.assertIsNone(result.name)
-        result = self.ts.add(cp)
-        self.assertIsNone(result.name)
-
-        ops = ['add', 'sub', 'mul', 'div', 'truediv', 'floordiv', 'mod', 'pow']
-        ops = ops + ['r' + op for op in ops]
-        for op in ops:
-            # names match, preserve
-            s = self.ts.copy()
-            result = getattr(s, op)(s)
-            self.assertEqual(result.name, self.ts.name)
-
-            # names don't match, don't preserve
-            cp = self.ts.copy()
-            cp.name = 'changed'
-            result = getattr(s, op)(cp)
-            self.assertIsNone(result.name)
-
     def test_comparisons(self):
         left = np.random.randn(10)
         right = np.random.randn(10)
@@ -1407,7 +1375,3 @@ class TestSeriesOperators(TestData, tm.TestCase):
         df['expected'] = df['date'] - df.index.to_series()
         df['result'] = df['date'] - df.index
         assert_series_equal(df['result'], df['expected'], check_names=False)
-
-    def test_scalarop_preserve_name(self):
-        result = self.ts * 2
-        self.assertEqual(result.name, self.ts.name)
